+++ src/packages/user/user-group/collection/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_USER_GROUP_COLLECTION_ALIAS = 'Umb.Collection.UserGroup';
