+++ src/packages/user/user-group/collection/index.ts
@@ -1,2 +1,2 @@
 export * from './repository/index.js';
-export { UMB_USER_GROUP_COLLECTION_ALIAS } from './manifests.js';
+export { UMB_USER_GROUP_COLLECTION_ALIAS } from './constants.js';
