+++ src/packages/user/user-group/collection/manifests.ts
@@ -4,9 +4,9 @@ import {
 } from './repository/manifests.js';
 import { manifests as collectionViewManifests } from './views/manifests.js';
 import { manifests as collectionActionManifests } from './action/manifests.js';
+import { UMB_USER_GROUP_COLLECTION_ALIAS } from './constants.js';
 import type { ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
 
-export const UMB_USER_GROUP_COLLECTION_ALIAS = 'Umb.Collection.UserGroup';
 const collection: ManifestTypes = {
 	type: 'collection',
 	alias: UMB_USER_GROUP_COLLECTION_ALIAS,
