+++ src/packages/user/user/repository/new-password/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_NEW_USER_PASSWORD_REPOSITORY_ALIAS = 'Umb.Repository.User.NewPassword';
