+++ src/packages/user/user/repository/new-password/index.ts
@@ -1,2 +1,2 @@
 export { UmbNewUserPasswordRepository } from './new-user-password.repository.js';
-export { UMB_NEW_USER_PASSWORD_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_NEW_USER_PASSWORD_REPOSITORY_ALIAS } from './constants.js';
