+++ src/packages/user/user/repository/unlock/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_UNLOCK_USER_REPOSITORY_ALIAS = 'Umb.Repository.User.Unlock';
