+++ src/packages/user/user/repository/unlock/index.ts
@@ -1,2 +1,2 @@
 export { UmbUnlockUserRepository } from './unlock-user.repository.js';
-export { UMB_UNLOCK_USER_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_UNLOCK_USER_REPOSITORY_ALIAS } from './constants.js';
