+++ src/packages/user/user/workspace/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_USER_WORKSPACE_ALIAS = 'Umb.Workspace.User';
